Added Image to logo editor converter
Added a vibe coded image editor that can convert images to logo data.

Renamed the Python folder to tools
