+++ tools/convert_logo.py
