+++ tools/osd_fonts/clarity.png
