+++ tools/osd_fonts/default.png
