+++ tools/osd_fonts/vision.png
